requirements.txt
django
djangorestframework
django-crispy-forms
python-slugify
python-dotenv
reverse
decorators
django-cors-headers